requirements.txt
openai>=1.60.0
httpx>=0.27.0
python-dotenv>=1.0.0
